+++ sdk/cosmos/azure-cosmos/azure/cosmos/__init__.py
@@ -21,6 +21,8 @@
 
 from ._version import VERSION
 from ._cosmos_responses import CosmosDict, CosmosList
+from ._embedding_provider import EmbeddingProvider
+from ._embedding_result import EmbeddingResult
 from ._retry_utility import ConnectionRetryPolicy
 from .container import ContainerProxy
 from .cosmos_client import CosmosClient
@@ -66,6 +68,8 @@ __all__ = (
     "ConnectionRetryPolicy",
     "ThroughputProperties",
     "CosmosDict",
-    "CosmosList"
+    "CosmosList",
+    "EmbeddingProvider",
+    "EmbeddingResult"
 )
 __version__ = VERSION